Rename all files according new schema of student learning resources
