+++ app/controllers/students/student_learning_resources_controller.rb
@@ -1,11 +1,11 @@
 module Students
-  class StudentsLearningResourcesController < ApplicationController
+  class StudentLearningResourcesController < ApplicationController
     before_action :authenticate_user!, :set_user
 
     def index
       # TODO: could be better y we have something like '@student.student_learning_resources' at this point
-      @student_learning_resources = @user.students_learning_resources.paginate(page: params[:page],
-                                                                               per_page: PER_PAGE)
+      @student_learning_resources = @user.student_learning_resources.paginate(page: params[:page],
+                                                                              per_page: PER_PAGE)
     end
 
     private
