+++ app/views/student_learning_resources/_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_with(model: students_learning_resource, url: teacher_students_learning_resources_url(@user)) do |form| %>
+<%= form_with(model: students_learning_resource, url: teacher_student_learning_resources_url(@user)) do |form| %>
 
   <div class="field is-horizontal">
     <div class="field-label is-normal">
