+++ app/views/students/student_learning_resources/_student_learning_resource.html.erb
