+++ app/views/students/student_learning_resources/index.html.erb
@@ -1,0 +1,16 @@
+<div class="container pt-5">
+  <h1 class="title">Your learning resources</h1>
+
+  <% @student_learning_resources.each do |student_learning_resource| %>
+    <%= render "students/student_learning_resources/student_learning_resource", learning_resource: student_learning_resource.learning_resource%>
+  <% end %>
+
+  <%= will_paginate @student_learning_resources, renderer: RemoteLinkPagination::LinkRenderer, class: "" %>
+  <br>
+  <div class="is-flex is-justify-content-flex-end">
+    <%= link_to(student_path(@user), class: "button is-info is-outlined mr-5") do %>
+      <i class="fas fa-undo"></i>
+      <span class="ml-2">Back</span>
+    <% end %>
+  </div>
+</div>
