--- app/views/students/students_learning_resources/index.html.erb
@@ -1,17 +1,0 @@
-
-<div class="container pt-5">
-  <h1 class="title">Your learning resources</h1>
-
-  <% @student_learning_resources.each do |students_learning_resource| %>
-    <%= render "teachers/learning_resources/learning_resource", learning_resource: students_learning_resource.learning_resource%>
-  <% end %>
-
-  <%= will_paginate @student_learning_resources, renderer: RemoteLinkPagination::LinkRenderer, :class=>"" %>
-  <br>
-  <div class="is-flex is-justify-content-flex-end">
-    <%= link_to(student_url(@user), class: "button is-info is-outlined mr-5") do %>
-      <i class="fas fa-undo"></i>
-      <span class="ml-2">Back</span>
-    <% end %>
-  </div>
-</div>
